requirements.txt
python-telegram-bot
psycopg2
pandas
python-dotenv
requests